examples/remote_agent_slim/requirements.txt
langgraph>=0.4.1
langchain>=0.3.25
langchain_openai>=0.3.21
langgraph-sdk>=0.1.53
langchain_experimental>=0.3.4
llama-index>=0.12.34
fastapi[standard]==0.115.11
requests>=2.32.3
starlette>=0.46.0
uvicorn>=0.34.0
packaging==24.2
python-dotenv>=1.0.1
pydantic>=2.10.6
pydantic-settings>=2.8.1
uvicorn>=0.34.0
requests>=2.32.3
python-json-logger>=3.3.0
slim-bindings>=1.0.0  # SLIM v1.x required
ioa_observe_sdk>=1.0.29
openai>=1.0.0
# If you need to debug langgraph studio
# debugpy==1.8.12
